Some Javadoc

+++ src/propra22/q8493367/metric/ManhattanDistance.java
@@ -2,8 +2,7 @@ package propra22.q8493367.metric;
 public class ManhattanDistance implements IMetric {
 
 	@Override
-	public int distance(int x1, int y1, int x2, int y2) {
+	public double distance(int x1, int y1, int x2, int y2) {
 		return Math.abs(x2 - x1) + Math.abs(y2 - y1);
 	}
-
 }
